change: move v1 chunk-snapshot types to openraft-legacy
# Summary

The v1 chunk-based snapshot surface — `InstallSnapshotRequest`,
`InstallSnapshotResponse`, `InstallSnapshotError`, `SnapshotMismatch`,
and `SnapshotSegmentId` — moves from the core crate into
`openraft_legacy::network_v1`, next to the v1 `RaftNetwork` trait that
is its only consumer. The old core paths still resolve, as empty
`#[deprecated]` marker types pointing to the legacy replacements, so
users get a guided migration window instead of an unexplained "cannot
find type" error.

# Details

Since chunk-based snapshot handling moved to the legacy adapter, core
merely defined these types without using them. Relocating them while
0.10 is still a breaking release keeps the v1 contract out of core for
good instead of carrying it through another major cycle. The
long-standing `TODO: remove` on `InstallSnapshotError` is resolved.

`openraft-legacy` gains a `serde` feature (forwarding to
`openraft/serde`) because the wire types now derive serde there. The
serialized representation is unchanged, so v1 peers interoperate across
this move.

`SnapshotResponse`, the v2 full-snapshot response, stays in core.

The deprecated stubs follow the existing tombstone pattern of the v1
`RaftNetwork` trait stub in `network/raft_network_v1.rs`: an empty item
at the old path, a **REMOVED** rustdoc header with a Migration section,
and a `deprecated` note naming the crate and the new path. Stubs are
added at every previously public location:
`openraft::raft::{InstallSnapshotRequest, InstallSnapshotResponse}`,
`openraft::errors::{InstallSnapshotError, SnapshotMismatch}`, and
`openraft::SnapshotSegmentId`. They carry no fields, impls, or serde
derives: they exist to be named in error messages and docs, not to be
used, and are expected to be removed in the release after 0.10. The
public-API test asserts the deprecated paths stay importable alongside
their legacy replacements.

The shared example alias file drops its two v1 aliases: most examples
do not depend on `openraft-legacy`, so the aliases would not resolve
there. The single-threaded example declares them locally instead.

Upgrade tip:

Replace core imports of the moved types with their legacy paths:

    use openraft_legacy::network_v1::InstallSnapshotRequest;
    use openraft_legacy::network_v1::InstallSnapshotError;

All of them are also exported via `openraft_legacy::prelude::*`.
