+++ examples/network-v1-http/Cargo.toml
@@ -16,7 +16,7 @@ repository = "https://github.com/databendlabs/openraft"
 
 [dependencies]
 openraft = { path = "../../openraft", features = ["serde", "type-alias"] }
-openraft-legacy = { path = "../../legacy" }
+openraft-legacy = { path = "../../legacy", features = ["serde"] }
 
 actix-web = { version = "4.0.0-rc.2", optional = true }
 reqwest    = { version = "0.12.5", features = ["json"] }
