+++ examples/network-v1-http/src/lib.rs
@@ -5,7 +5,6 @@ use openraft::BasicNode;
 use openraft::OptionalSend;
 use openraft::RaftTypeConfig;
 use openraft::errors::Infallible;
-use openraft::errors::InstallSnapshotError;
 use openraft::errors::NetworkError;
 use openraft::errors::RPCError;
 use openraft::errors::RaftError;
@@ -15,8 +14,6 @@ use openraft::network::RPCOption;
 use openraft::network::RaftNetworkFactory;
 use openraft::raft::AppendEntriesRequest;
 use openraft::raft::AppendEntriesResponse;
-use openraft::raft::InstallSnapshotRequest;
-use openraft::raft::InstallSnapshotResponse;
 use openraft::raft::VoteRequest;
 use openraft::raft::VoteResponse;
 use openraft_legacy::prelude::*;
@@ -35,15 +32,15 @@ pub mod actix {
     use openraft::Raft;
     use openraft::RaftTypeConfig;
     use openraft::errors::Infallible;
-    use openraft::errors::InstallSnapshotError;
     use openraft::errors::decompose::DecomposeResult;
     use openraft::raft::AppendEntriesRequest;
     use openraft::raft::AppendEntriesResponse;
-    use openraft::raft::InstallSnapshotRequest;
-    use openraft::raft::InstallSnapshotResponse;
     use openraft::raft::VoteRequest;
     use openraft::raft::VoteResponse;
     use openraft::storage::RaftStateMachine;
+    use openraft_legacy::network_v1::InstallSnapshotError;
+    use openraft_legacy::network_v1::InstallSnapshotRequest;
+    use openraft_legacy::network_v1::InstallSnapshotResponse;
     use openraft_legacy::prelude::ChunkedSnapshotReceiver;
     use serde::Serialize;
     use serde::de::DeserializeOwned;
