+++ examples/raft-kv-memstore-single-threaded/Cargo.toml
@@ -17,7 +17,7 @@ repository = "https://github.com/databendlabs/openraft"
 
 [dependencies]
 openraft           = { path = "../../openraft", features = ["serde", "single-threaded", "type-alias"] }
-openraft-legacy = { path = "../../legacy" }
+openraft-legacy = { path = "../../legacy", features = ["serde"] }
 
 futures            = { version = "0.3" }
 serde              = { version = "1.0.114", features = ["derive"] }
