+++ openraft/src/raft/message/mod.rs
@@ -19,7 +19,9 @@ pub use append_entries_request::AppendEntriesRequest;
 pub use append_entries_response::AppendEntriesResponse;
 pub use client_write::ClientWriteResponse;
 pub use client_write::ClientWriteResult;
+#[allow(deprecated)]
 pub use install_snapshot::InstallSnapshotRequest;
+#[allow(deprecated)]
 pub use install_snapshot::InstallSnapshotResponse;
 pub use install_snapshot::SnapshotResponse;
 pub use log_segment::LogSegment;
